refactor: split controls model from screen
